backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
cachetools==5.5.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
cloudinary==1.43.0
colorama==0.4.6
Deprecated==1.2.18
dj-database-url==2.3.0
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
googletrans==4.0.0rc1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.0
pydantic_core==2.33.0
PyJWT==2.9.0
pyparsing==3.2.3
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==4.3.4
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wrapt==1.17.2
